temp patch to !purge to make it so that non-mods cant abuse the command
it now currently only recognizes !purges from me, will change later to accept !purge from all mods

+++ bot.js
@@ -18,8 +18,8 @@ client.on('message', async message => {
 
   // Bot listens to messages with the prefix !
   if (message.content.substring(0, 1) == '!') {
-    const args = message.content.slice(1).trim().split(/ +/g);
-    const command = args.shift().toLowerCase();
+    const args = message.content.slice(1).trim().split(/ +/g); // removes the prefix, then the spaces, then splits into array
+    const command = args.shift().toLowerCase(); // removes the command from the args array
 
     // Commands
     switch (command) {
@@ -74,16 +74,18 @@ client.on('message', async message => {
         break;
 
       case 'purge': // This command removes all messages from all users in the channel, up to 100.
-        // get the delete count, as an actual number.
-        const deleteCount = parseInt(args[0], 10);
+        if (message.author.id == '355534246439419904') {
+          // get the delete count, as an actual number.
+          const deleteCount = parseInt(args[0], 10);
 
-        if(!deleteCount || deleteCount < 2 || deleteCount > 100)
-          return message.reply("Please provide a number between 2 and 100 for the number of messages to delete");
+          if(!deleteCount || deleteCount < 2 || deleteCount > 100)
+            return message.reply("Please provide a number between 2 and 100 for the number of messages to delete");
 
-        // So we get our messages, and delete them. Simple enough, right?
-        const fetched = await message.channel.messages.fetch({limit: deleteCount});
-        message.channel.bulkDelete(fetched)
-          .catch(error => message.reply(`Couldn't delete messages because of: ${error}`));
+          // So we get our messages, and delete them. Simple enough, right?
+          const fetched = await message.channel.messages.fetch({limit: deleteCount});
+          message.channel.bulkDelete(fetched)
+            .catch(error => message.reply(`Couldn't delete messages because of: ${error}`));
+        }
     }
   }
 });
